Update #59 - Contact Birthday
Given a Contact with the Birthdate field set to some date, return true if today is the Contact's birthday and return false if not. Assume that a future date will not be set on the Birthdate field.

Given the following test code:

Contact c1 = new Contact();
c1.Birthdate = Date.newInstance(1992, 5, 15)
The expression isBirthday(c1) should return true if executed on 5/15/2022 or 5/15/2020.

HINT1:
You can obtain today's date using the System.today() method
HINT 2:
Consider using the month and day methods on the Date class. See documentation for the Date class.
HINT 3:
Obtain the month and day for the Contact's birthdate and today's date. Compare them to see if today is the Contact's birthday.