Create #83 - Employee Email
Given a string employeeId a map of string to string employeeIdToEmail that contains employee IDs as keys and the employee's email address as the value, return the email associated with the passed employeeId. If the employee ID is not found, return 'info@apexsandbox.io'.

For example, if the map employeeIdToEmail contains the following keys and values:

'6752' -> 'cooper@corporation.com'
'19228' -> 'umair@universalcontainers.com'
'4823' -> 'alicia@acmeenterprises.com'
The method call employeeEmail(employeeIdToEmail, '4823') should return 'alicia@acmeenterprises.com'.
HINT 1:
The get method in the Map class allows you to obtain a value from a key, and the containsKey method allows checking if a key exists.
HINT 2:
Check if the map contains the employee ID using the containsKey method. If it contains the key, pass the employeeId into the get method to obtain the email and return it. Otherwise, return the default email.