Update #2 - Largest Element
Implement the method findLargest that takes as input a non-empty list of Integers nums, and returns the largest Integer in the list.

Example: findLargest(new List {5, 2, 8, 4, 2, 1}) evaluates to 8 because 8 is the largest Integer in the list.
HINT1:
Use the .size() function to obtain the size of the input list
HINT 2:
Use the [] notation or the .get(index) method to obtain the element at index.
HINT 3:
Declare an Integer to hold the largest value and initialize it with the element at index 0 in the list. Use a for loop to check if each element in the list is the largest encountered so far.