Create #71 - Positive Integers
HINT 1:
Remember to not modify the input list. The returned list should be a new list.
HINT 2:
Use new List() to create a new empty list and use the add(listElement) method to add new elements into the list. Refer to documentation for the List class.
HINT3:
Start by creating an empty list of integers using the list List positives = new List();. Loop over numbers and if the number is greater than 0, add the number into the new list with the line positives.add(numbers[i]);. When finished, return positives.