Create #73 - Sorted List
A list is considered to be sorted ascending when no element in the list is smaller than the preceding element if one is present. Similarly, a list is considered sorted descending if no element in the list is larger than the preceding element if any.

Implement the method isSorted that takes as input a list of integers numbers, returns true if the list is sorted in any direction (acsending or descending), and returns false otherwise.

Example: isSorted(new List<Integer> {5, 2, 0, -1}) evaluates to true because the input list is sorted descending.
HINT1
Lists with less than 3 elements are always sorted.
HINT2
Loop over the entire list to see if either the ascending or descending condition is broken by any element. If at least one of the two conditions is unbroken throughout the array, the list is sorted.
HINT3
For a more efficient approach, start by looping over the list until you encounter an element that differs from the previous one.
HINT4
Once you have encountered a different element, determine which sort order the rest of the list should maintain. Continue looping and ensuring the sort order is maintained and return false as soon as it is broken.